MB-47197: Add support for cross-connection limits [1/n]
As part of MB-46881 one needs to be able to rate limit clients
based upon total usage across all their connection.

In order to try out the prototype set the environment
variable COUCHBASE_DBAAS before starting memcached.

The default user limits is stored in JSON files in

    <root>/etc/couchbase/kv/security/user.d/default.json

(There is a readme in the directory describing the format
and how to add per-user overrides)

In the current prototype the file is loaded _the first time_
the user logs in. Until the file has been loaded (or there
is no file for the user) the user is assigned the following
limits:

    Network ingress: 10MB/min
    Network egress : 10MB/min
    Operations     : 60/min
    Connections:   : 1 concurrent

We don't need the metrics to be absolute correct, so we use a
"sloppy" mechanism to enforce the limitations (which means
that one may exceed the limitations with a few ops/bytes, but
we don't care that much).

The check use a "SloppyGauge" which holds a minute-counter and
a value. Whenever we detect that we're operating in a different
minute we reset the counter. This isn't done with a lock (just
two atomics) which means that multiple threads may operate on
the two variables "at the same time" (two threads may decide
to clear the value and thread 1 clears it and sets its new
value before thread 2 clears it again). We can accept that
limited error margin to avoid having to add locks.

While the patch schedule a task to run ever 30 minute to
purge idle tenants none of them gets purged as the method to
decide if it should be purged isn't implemented yet.

One may receive cross-connection tenant stats by requesting
the tenant stat group:

    mcstat -b bucket tenants

Which retrieves the statistics for all tenants, or one may
receive for a single tenant by using:

    mcstat -b bucket 'tenants {"domain":"local","user":"jones"}'

The system currently only collect the following information
on a per tenant base:

    {
      "connections": {
        "current": 10,
        "total": 12
      },
      "constrains": {
        "connections": 10,
        "egress": 52428800,
        "ingress": 52428800,
        "operations": 6000
      },
      "egress": 144552,
      "ingress": 144264,
      "operations": 6021,
      "rate_limited": {
        "connections": 1,
        "egress": 0,
        "ingress": 0,
        "operations": 1
      }
    }

Enforcing concurrent connections is a bit "hard" as we don't
know the tenant until the connection is successfully authenticated.
We could abort the authentication and reply with "rate limit"
instead of authentication success, but that could potentially
break (old) clients. We don't want to return success and just
close the connection, as we want to be able to return an error
message letting the client know why it was disconnected.
To work around the problem we're disconnecting the clients whenever
they try to execute packets. That means that you could in theory
connect way too many connections to the server, but you wouldn't
be able to use any of them as they would be disconnected when
they try to execute a packet.

Add an entry to the audit trail when a tenant gets rate limited
to make it easier to monitor when; where and why.

Change-Id: If4f063bff7efade8cf6e8afd611f04cb26a801bf
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147171
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>