Revert "MB-43373: Abort memcached if exception thrown from GlobalTask::run v2"
Now FollyExecutorPool no longer catches exceptions on bg threads (see
MB-44253), we don't want to catch them ourselves in GlobalTask - we
want them to be uncaught and trigger std::terminate with a full
backtrace to where the exception was thrown from.

(Note: This was partially reverted by 8828aa7b52 ("Revert "MB-43818:
GlobalTask::execute: Log task info if exception thrown"), but the
throw was not removed. This patch completes the revert.

This reverts commit b6ed8d0f9c8f55a8ee7a3b19a4e1e307d5459eeb.

Change-Id: I4b4085ee4b4eae3e397d79ade51cb3fc88f67e87
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147330
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>