MB-41257: Don't enable traffic if Warmup can't write vb state to disk
Currently during warmup we will try and persist a new vbucket state to
disk during Warmup::populateVBucketMap(). This is to ensure we enqueue
any new failover uuid in the checkpoint manager, before we add the
vbucket to vbucket map (allowing us to accept DCP connections). So its
the first thing that written to disk. This gives us a good place to
ensure that we can write to the files of a vbucket or not, to ensure
our storage is in a good state. However, currently we don't take
advantage of this. Waiting for the flusher task to persist the state
to disk without reporting its success.

To help ensure we don't write to vbuckets that we can't reliably write
to. We have modified the Warmup::populateVBucketMap() to call the flush
of the vbucket state to disk immediately after enqueuing it in the
checkpoint manager. If the flush fails, we will set a flag to disable
the ability to enable traffic. This is to ensure we don't accept writes
that we will be unable to persist to disk. But will allow clients to
perform gets effectively putting the node a read only state.

Change-Id: Ibc83fa850ea08b4ab10c32953f2225c0bdfd5753
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147349
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>