MB-44674: Disable bufferevent space optimization for TLS
The buffer optimization added to reduce the memory fragmentation
(and extra allocations) by using evbuffer_reserve_space and
evbuffer_commit_space introduced a 30% drop in throughput when
using TLS.

It is not fully clear _why_ we see a drop (without TLS we're running
at a higher ops/sec and there is no drop). That indicates that it is
related to OpenSSL and encoding the TLS frames.

For now just disable the optimization for TLS while performing deeper
analysis.

Change-Id: I537b7ec3554554204deb901994c56c15469cadcb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147419
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>