MB-44562: Queue producers/BFMs when we hit running backfill limit
Currently we have a maximum limit of 4096 concurrently running
backfills. This limit is variable depending on max data size and
may be lower. Should we hit this limit, backfills are added to the
pending queues of their BackfillManagers which will check the limit
and attempt to move them to the initializing queue each time they
run a backfill. Should a BackfillManager have no running backfills
then it will snooze for a set period of time. If a BackfillManger
does have running backfills it will yield for other tasks and run
again as soon as possible. As a BackfillManager tries to move
pending backfills on every run, regardless of how many backfills it
has currently running, we are unfairly giving the BackfillManager
with more backfills a higher chance of getting a new one. This can
lead to a resource starvation for some BackfillManagers/Producers.

Reduce the chance of this resource starvation by queueing
BackfillManagers for the next backfill resource when we hit the
maxRunning backfill limit.

Change-Id: I10528570bad0314c3ae84ebc78a0911ec6193267
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147424
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>