MB-37520: Change DcpProducer::StreamsMap to AtomicHashArray size 1024
DcpProducer::StreamsMap is implemented using folly::AtomicHashMap (AHMap),
which uses a linked list of fixed size AtomicHashArray chunks to
support arbitrary map sizes.

While the performance of AHMap::find() is very good (typically
single-digit nanoseconds) if the maximum size of the map is correctly
predicted up front, it falls off drastically if the size exceeds that,
as it must

a) Scan potentially the entire initial subMap (AtomicHashArray), and
b) Perform scans in all subsequent subMaps.

This cost is made worse by a feature of our key & hash function which
is a uint16_t with the identity function. This works very well if size
< expected size, as there should be virtually no collisions, but if
size > expected size then collisions will occur, and given many
consequtive slots are likely to already be full (consequative Vbids)
then many collections can occur.

In numbers, this looks like the following, where we have a
AtomicHashMap with expected size 512 (i.e. largest expected number of
active VBs on a production cluster of at least 2 nodes):

    Running ./ep_engine_benchmarks
    Run on (12 X 2900 MHz CPU s)
    CPU Caches:
      L1 Data 32 KiB (x6)
      L1 Instruction 32 KiB (x6)
      L2 Unified 256 KiB (x6)
      L3 Unified 12288 KiB (x1)
    Load Average: 2.69, 3.40, 6.90
    ----------------------------------------------------------------------------------------------------
    Benchmark                                          Time             CPU   Iterations UserCounters...
    ----------------------------------------------------------------------------------------------------
    DcpProducerStreamsMapBench/findMap/128          3.46 ns         3.46 ns    181408716 capacity=640 numSubMaps=1
    DcpProducerStreamsMapBench/findMap/256          3.35 ns         3.35 ns    198131338 capacity=640 numSubMaps=1
    DcpProducerStreamsMapBench/findMap/384          3.42 ns         3.42 ns    213520744 capacity=640 numSubMaps=1
    DcpProducerStreamsMapBench/findMap/512          3.29 ns         3.29 ns    204302017 capacity=640 numSubMaps=1
    DcpProducerStreamsMapBench/findMap/640          3.26 ns         3.25 ns    212053098 capacity=640 numSubMaps=1
    DcpProducerStreamsMapBench/findMap/768           157 ns          157 ns      4574954 capacity=800 numSubMaps=2
    DcpProducerStreamsMapBench/findMap/896           288 ns          288 ns      2435325 capacity=991 numSubMaps=3
    DcpProducerStreamsMapBench/findMap/1024          413 ns          413 ns      1709573 capacity=1.221k numSubMaps=4

Note the dramatic drop-off once the size exceeds the expected size of
512; particulary once we get up to 1024 elements - Time is around 100x
worse.

While >512 vBuckets would not be expected to be an issue (would only
affect a single-node cluster / 2-node cluster after failover), in the
case of MB-37520 we are measuring DCP backfill peformance of a single
node. As such, we do hit this situation.

To address this particular use-case, increase the expected size to
1024 (technically max_vbuckets) - this means we always see the
single-digit ns performance.  Additionally, given we no longer need
the resize capability we can switch to the simpler AtomicHashArray -
i.e. the type used for each sub-map of AHMap. This avoids some
overhead in managing the different sub-maps, and gives even better
performance - consistenly under 2ns:

    Running ./ep_engine_benchmarks
    Run on (12 X 2900 MHz CPU s)
    CPU Caches:
      L1 Data 32 KiB (x6)
      L1 Instruction 32 KiB (x6)
      L2 Unified 256 KiB (x6)
      L3 Unified 12288 KiB (x1)
    Load Average: 2.69, 3.40, 6.90
    ----------------------------------------------------------------------------------------------------
    Benchmark                                          Time             CPU   Iterations UserCounters...
    ----------------------------------------------------------------------------------------------------
    DcpProducerStreamsMapBench/findArray/128        1.72 ns         1.72 ns    403732798 capacity=1024
    DcpProducerStreamsMapBench/findArray/256        1.73 ns         1.73 ns    404257408 capacity=1024
    DcpProducerStreamsMapBench/findArray/384        1.71 ns         1.71 ns    411288096 capacity=1024
    DcpProducerStreamsMapBench/findArray/512        1.72 ns         1.72 ns    415773249 capacity=1024
    DcpProducerStreamsMapBench/findArray/640        1.72 ns         1.72 ns    408391870 capacity=1024
    DcpProducerStreamsMapBench/findArray/768        1.68 ns         1.68 ns    420964008 capacity=1024
    DcpProducerStreamsMapBench/findArray/896        1.70 ns         1.70 ns    413403730 capacity=1024
    DcpProducerStreamsMapBench/findArray/1024       1.68 ns         1.68 ns    428826974 capacity=1024

An AtomicHashArray of 1024 elements requires 24,656 Bytes of memory
(1024 x std::pair<uint16_t, StreamMapValue>). This is compared to
15,554 Bytes for a AtomicHashMap of expected size 512 - a 9kB increase
per DcpProducer. This seems to be a reasonable trade-off given the
improved, consistent performance and no risk of severe degredation if
size > expected_size.

Change-Id: I68937bd851a01a70e611e1d7e0af794e79f1d7f6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147606
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>