MB-41257: Abort warmup if we're unable to get collection states
KVStore::getCollectionStats() can fail to get collection stats for other
than the reason of NOT_FOUND. However, currently the KVStore
implementations treat any error in the same way a NOT_FOUND. This can
cause us to Warmup with incorrect collection stats. If
getCollectionStats() fails to get stats for any other reason than
NOT_FOUND we should fail Warmup as otherwise we in an inconsistent
state. Furthermore, this could be indication that the storage is failing
or files on disk are corrupted.

To allow us to abort warmup if getCollectionStats() fails,
getCollectionStats() returns a std::pair<> with a bool to indicate
successful fetching of collections stats.

Its worth noting as of writing this patch, Magma gets its
collection stats in memory so won't fail if the data files are corrupted
on disk. So if makeFileHandle() passes getCollectionStats() will
succeed. This is not the case for Couchstore.

This patch also adds tests for KVStore::getCollectionStats(). But does
not add new Warmup tests as there's not way of blocking in
Warmup::loadCollectionStatsForShard() between the call to
KVStore::makeFileHandle() and KVStore::getCollectionStats().

Change-Id: Id10076c6c87981ae793f83be584dae7fb40b6ff3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147724
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>