MB-44590: Improve compaction logging
Recent changes to compaction scheduling moved the compaction
created logging to the constructor of the CompactTask. This
lost logging when compaction is rescheduled, i.e. a log may
contain (simplified output).

Compaction created
Compaction completed
Compaction completed

This commit updates the logging so that the creation or rescheduling
is logged, along with the delay value and the config (considering
that reschedule may merge configs).

Compaction created with delay 0s
Compaction completed
Compaction rescheduled with delay 0s
Compaction completed

Change-Id: Icefa836f3790a60de89d16abd1f17f0a022994a8
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147709
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>