MB-44590: Maintain dropped collections from replayPrecommitHook
Prior to this fix if the replay function was used (because flush and
compaction interleaved) the dropped collections meta-data wasn't being
maintained. The associated MB identified a case where the meta-data
stored a dropped collection which had been fully purged.

This fix updates replayPrecommitHook so that the dropped collection
meta-data is maintained. The new code has to consider that new
collections could of been dropped during the 'interleave', this means
that the final output is no longer just the removal of the  dropped
collection data, but is either deleted or just the newly dropped
collections (so any following full compaction can purge the new dropped
collection).

Change-Id: I9eeb65c8cbaa195ec451087439fca5ddf9818fe8
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147364
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>