MB-44590: ScanContext startSeqno should not always be 0
Noted as part of MB-44694 an exception message that printed the
ScanContext showed the startSeqno was 0. This was incorrect and
down to the default initialisation of 0 and later use of std::min.

With this change we now correctly set the startSeqno so that the
seqno range check now correctly skips items not in the dropped
range.

Change-Id: I855f0a21975dfecc2e592f75395ab19cb842c213
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147710
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>