MB-45378: Statically link all engines
Two issues[1,2] have so far been identified with Folly when multiple
instances of Folly end up within the same process - i.e. multiple
executables / DSOs present in memcached process.

Both issues can be fixed by changing the visibility of certain folly
symbols, but our current usage model (linking into shared libraries
with visibility=hidden) is not the common model for Folly and hence the
concern that this isn't a stable or safe way to continue to consume
Folly.

As such, move as much as possible of our linking to be static -
specifically any library which uses Folly should be linked statically.

This patch does this for all engines, so they are part of the
memcached binary.

[1]: https://github.com/facebook/folly/issues/1431
[2]: https://github.com/facebook/folly/issues/1558

Change-Id: Ic9633db89e9e769146e7869a5122b6720ea2e1eb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147715
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>