MB-41257: Reset vbucket to avoid bad state of PDM
Harden rollback for a rare case where the rollback of persisted seqnos
succeed but loadPreparedSyncWrites() fails. In this case We should
reset the vbucket and re-stream from seqno 0. Otherwise we will have
rolled back to the rollback seqno but may not have correctly loaded
prepares from disk which could cause us to be an inconsistent state
with the DCP producer.

This patch also makes another modification to log if the call to
doRollback() fails. The patch also adds a unittest using GMOCK to fail
the call to loadPreparedSyncWrites() due to a read error. Thus, ensuring
that we roll back to zero when loadPreparedSyncWrites() fails.

Change-Id: I15e53673c15dd64fe659ec649a1e187f4585a982
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147744
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>