MB-44539: More efficient generation of collection stats document ID
Profiling a workload which was writing to 1k collections and it  was
noted that CollectionID::to_string was relatively high for the writer
threads (as part of the collection stat update).  This function was
originally intended for debug use but is used in the generation of the
local document ID, and it's not the most efficient way to generate
0x<cid>.

For CouchKVStore a new method is added which uses fmt to generate the
entire document name in one call. With this change, the generation of
document names has disappeared from the profile.

Additionally the CollectionID::to_string is updated to use the more
efficient formatter.

Change-Id: I40707f5ac8a9403a6a94345168a35b7e9d9520b8
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147770
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>