MB-44168: Handle disk-write failures at DB bootstrap
At vbucket creation we generate a new couchstore file. That operation
may fail (eg IO failures) and leave a corrupted file on disk. From the
couchstore's PoV even an empty file is corrupted, as the storage will
not be able to open and work on ever again.

With this patch we introduce a bootstrap procedure that creates new
couchstore's file atomically, and we execute it when the system detects
that there is no file on disk yet for a vbucket.

The result of the new bootstrap procedure is either a new well-formed
couchstore file that contains the first header at filepos 0, or no file
at all.

Change-Id: I9d5c907b886b54c329193c12f7dc99c354c71bff
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147938
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>