MB-44539: Don't read dropped collections metadata unless it exists
Profiling the flusher shows that "getDroppedCollections" called
from saveDocs accounted for 2.2% to 2.5% of the flusher threads
profile, and this is for a workload where the document didn't
actually exist, couchstore still does some work to read and
decompress data to find that the document is not present.

The dropped collection data is only needed by the flusher for
stats accounting when it exists and we already have some state
in the VB::Manifest which indicates if this document exists.

This commit makes that VB::Manifest 'bool' readable from
CouchKVStore::saveDocs so that the call to getDroppedCollections
can be skipped unless needed.

This also requires some work in the compaction code as compaction
is the entity which may finally delete the document (once compaction
has ran and purged all dropped collections).

Change-Id: Id7274161a07bf6f59b2bcc6b7757945fa1546a79
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147911
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>