MB-44807: get_collection|scope_id: allow 'path' input as key or value
The max scope name + collection name exceeds the limit of the
protocol key field. This commit changes both get_scope_id and
get_collection_id so that the input can be encoded in the key
or the value.

This is temporary to alleviate migration to updated clients,
long term only the input via value will be accepted.

Change-Id: I609bbb9c3ba4ebd3c07d6019523b300d0de79f63
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148066
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>