Add back copyToOutputStream(gsl::span<std::string_view> data)
This patch adds back the refactoring of the call chain to
allow for optimizing the buffer use in
http://review.couchbase.org/c/kv_engine/+/146916
(which unfortunately caused multiple threads to read/write
from the underlying buffer without keeping it locked).

The refactor introduced in the patch was however useful for
trying to reduce the number of explicit calls to copyToOutput
stream so that it'll make it easier for us in the future
to do a cross-buffer optimization.

Change-Id: Ifc704882d3f287e7805a6bdfb72cec22f4f0be63
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148171
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>