MB-44858: Correct stats if updated between cursors and flush fails
Currently if we update an item in an open checkpoint in the
checkpoint manager between the backup persistence cursor and the
regular persistence cursor then we will overcount the stats if the
flush fails. Correct this by tracking a set of stats that are
set when we add an item between persistence cursors, reset when a
flush completes successfully, and applied when a flush fails.

Change-Id: Id441ce37d67e1e6c222573e7eee151fc4aba41d3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148286
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>