MB-38007: Limit the DCP send queue from step() to 1MB
There is a 60% increase in rebalance time for ephemeral
bucket with the move to bufferevents and the way we structured
the buffering for sending DCP messages.

In the initial implementation we allowed up to 40MB of data to
prepared for sending to the other side, before allowing bufferevent
to send the data in bigger chunks. From the looks of it one should
think that it should be more efficient (fewer system calls, allowing
for bigger TLS frames/bigger IP packets etc), but the reality is
different. It could be that the buffer size don't align very well
to the underlying flow control in ep-engine causing us to do full
stops sending data and then wait for the other side to process the
entire chunk before it returns the flow control message to start it
again and a new blob would be sent and a full stop occurs etc.

This patch set the buffering to 1MB which brings us back in the
ballpark from before the move to bufferevent and we may try to
improve this area later.

Change-Id: I1d92951889f48e4eb60748604962fbcf32ece489
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148350
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>