MB-36610: Use runInEventBaseThread from the external auth mgr
The external authentication manager operates in its own thread
context, but it needs to send data on connections bound to
different threads. Previously we would try to lock the libevent
instance the connection was bound to in order to make sure we
wouldn't interfer with the connection. Now that we're using
folly::EventBase we can just use runInEventBaseThread to do
the same. Note: We've got a reference count to the other
connection so that it cannot disappear while we're trying to
use it (and hence we schedule the decrement of the reference
counter to be run in the event base thread when we want to
release it)

Change-Id: I958659b4f05c1124f02c79ba20490b9d43a6b848
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148532
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>