MB-41639: Pass storage thread setting directly to engine
Currently we can supply the storage backend with a thread number
configuration parameter (storage threads) via the process wide
memcached config. The only storage backend that uses this
configuration parameter is magma which registers a callbck to be
notified of this change via the server api. This is problematic as
we don't currently unregister this callback should the magma bucket
be deleted. Unfortuantely unregistering the callback in this case
does not solve the problem as we have a cardinality issue (1 memcached/
server api instance/ multiple buckets). Were we to unregister the
callback we could miss notifications required for other magma buckets.

Remove the callback logic from the server api and instead pass the
configuration parameter down to each bucket when it changes so that
the bucket can choose to either act on it or ignore it. This means that
magma will attempt to change the number of threads for each magma bucket
that exists, but the configuration will be the same and the net result
will just be extra log messages (which could be moved/removed).

Change-Id: Ia0a73ac43af34c6e2864c116336ce811707d6c89
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148552
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>