MB-44832: Allow abort with prepare seqno < snap start at backfill
Currently we disconnect if we receive an abort in a backfill
snapshot that has a prepare seqno lower than the snapshot start
seqno. This sanity check was added under the assumption that we
must have received the prepare already but that may not be the case
if we disconnect mid snapshot or roll back the portion of the
snapshot containing the abort. In this case we will logically be
receiving a "catchup" backfill snapshot with snap start != 0 and
the prepare seqno of an abort could be lower than the snap start.

Change-Id: I112765d151c747fab2d365b904b09dcdf2015bdb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148833
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>