MB-44665: Remove deadlock in ADM when catching exception
New debug logging was added to the ADM for MB-41434, to understand
why ActiveDurabilityMonitor::State::updateHighCompletedSeqno() was
throwing due to an invariant failure. But this introduced a deadlock
situation where if updateHighCompletedSeqno() we would catch the thrown
exception and log it. As part of the logging we also want to dump the
ADM's state but this requires a read lock on the ADM's state. Which
can't be taken as we've already taken a write lock in the caller of
updateHighCompletedSeqno().

To fix this access each member variable directly as if we're in
updateHighCompletedSeqno() we must have a write lock on the ADM's state.

Change-Id: I17130580f54a7efa4b39eb68a1a2c855c20dee89
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148921
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>