MB-36956: Increase upper bound on Reader/Writer/Shard count to 128
Now we are using FollyExecutorPool for our thread pools and
scheduling, we don't have the same contention in the various task
queues that CB3ExecutorPool had. Additionally, we have some perf
machines which exceed 64 logical cores per node (e.g. aether, athena,
hemera, hercules all have 80 logical cores), and hence we are
significantly constraining the performance they are capable of.

Increase the limit to 128 in each instance.

Note: We _could_ remove the upper bound entirely, and just trust what
the system / user specifies; but a gradual increase of the limit seems
a sensible approach.

Change-Id: Ie4745711d7dd13139beedf88595576908d0d8256
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148970
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>