MB-37520: Fix incorrect VBReadyQueue map_size stat
As seen in DCP statistics, the map_size stat is incorrectly reporting
the total capacity of the map, not the number of items it contains:

 eq_dcpq:<<connection>>:dcp_ready_queue_contents:
 eq_dcpq:<<connection>>:dcp_ready_queue_map_contents:
 eq_dcpq:<<connection>>:dcp_ready_queue_map_size:                1024
 eq_dcpq:<<connection>>:dcp_ready_queue_size:                    0

This is due to the switch to boost::dynamic_bitset which uses size()
to return the capacity, not the number of items currently set. Change
to use count() to fix.

Change-Id: I98c350599397f606d6e101252973618f2ae70fc7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/148979
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>