Don't reset cursors if no checkpoint created in CM::createSnapshot
CM::resetCursor repositions cursors at CheckpointList::begin and updates
the accounting of cursor in checkpoints after a CheckpointList change.

There is no reason for making the call in code paths that don't modify
the CheckpointList.

The call removed was the only one passing 'resetPersistenceCursor =
false' at resetCursor(), so the argument is removed.

Change-Id: I987498742426eb4f2b4939faf7c4509194df6783
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149375
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>