MB-45211: Ensure FollyExecutorPool updates waketime
When a task wishes to be immediately run again (e.g., when momentarily
"yielding" to avoid monopolising a thread for a very long task) it can
return true without calling snooze().

In this case, the pool should update the waketime to "now". If it does
not, the calculated scheduling delay will be derived from the old
waketime value from _before the task was run_, and will be much higher
than it should be.

Now when immediately rescheduling a task, the FollyExecutorPool updates
the waketime (consistent with the CB3 behaviour).

Change-Id: Ia104c0a4349ab2f2725a295a7342044e44d9b94b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149549
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>