MB-45206: Fix logging in Connection::event_callback
In memcached a clean connection shutdown by the peer is caught as
BEV_EVENT_EOF.

But, bufferevent may generate composed events like BEV_EVENT_EOF |
BEV_EVENT_ERROR | BEV_EVENT_WRITING, whose interpretation is "error at
writing on a closed socket".

In that case we are currently logging the event as an error, while the
"error" part is just a consequence of a clean shutdown.

Fix logging in Connection::event_callback accordingly.

Change-Id: I5dac665e46c31b1e8725816b0066810407977bb0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149603
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>