MB-45206: Log errno in Connection::event_callback
Even with BEV_OPT_DEFER_CALLBACKS (which is our default bufferevent
setting) the library ensures that errno is consistent with the event
passed in input to the user-defined callback (ie,
Connection::event_callback in our case).

That means that we can safely dump errno when we get a BEV_EVENT_ERROR
and trust what it reports.

See the libevent code for details.

Change-Id: I98d808028259b0fbb74b6e6bdfbf71df96a55c98
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149627
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>