Fix CouchstoreFileAccessGuard() on Windows
Currently CouchstoreFileAccessGuard() doesn't work as intended when
used in DenyAll mode. This is due to the fact that chmod() on Windows
works by modifying the files attributes, but on Windows there is no
attribute that states if a user can read a file only an attribute to
make it read-only. Instead Windows implements this functionality using
ACL's (Access control lists) however, to access these they require
verbose code. So instead simply create an exclusive handle to the file
for the lifetime of the CouchstoreFileAccessGuard() when in the DenyAll
mode.

Change-Id: I3cdb8e99c021135ca8121f430df917a2a7e8b85f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149757
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>