MB-45221: 3/n Split Collections::Flush into two classes
To assist with resolving MB-45221 we need to re-use the
accounting code of Collections::VB::Flush inside compaction
replay. The idea is that as we 'copy' items during the replay
the accounting code can be re-used to account for the changes
made to the new database.

This commit aims to be a code move only, splitting the code
and data which can be shared between a flush and replay into

  Collections::VB::FlushAccounting

This new class owns only the map of collection stats and maps
of dropped collections.

The existing Collections::VB::Flush class has the new class
as a data member.

Change-Id: I9fe6c409253775320f412106b46818d736ceafcc
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150012
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>