MB-45505: Update vbuckets with collections earlier in warmup
Warmup has code which will ensure that active vbuckets
are equal with the bucket's view collections. This exists
so that if KV has responded success to set_collections
but then warmed-up, vbuckets don't come back with old
collection state.

However this code should be earlier in warmup, before the
vbucket is added to the vbmap. This makes it much easier
to reason about when considering various failure scenarios.

For example in the previous implementation, we may of signalled
success to set_collections and dropped a collection. But then a
window exists in warmup where that collection maybe visible. It
is simpler to reason about failure (especially quorum loss) if
the collection state is synced whilst the vbucket is not visible.

Change-Id: I432ec5b401cd240cd36d5e454a17ff17df3934e2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149996
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>