MB-45255: Skip backfillMgr update if null
There is a race condition on DcpProducer during bucket shutdown if a
new stream request arrives at the "wrong" time. This causes
DcpProducer::backfillMgr being dereferenced when it is null, resulting
in a segfault:

1. Start with 1 producer and an existing stream (so DcpProducer is not
   paused).
2. DcpProducer::streamRequest starts, succeeds but _doesn’t get as far as
   adding to DCP conn map.
3. DcpConnMap::closeStreams (due to bucket shutdown)
   1. closes all streams for each producer via DcpProducer::closeAllStreams
   2. which resets backfillMgr ptr.
4. DcpProducer::streamRequest continues; adds new stream into map.
5. DcpConnMap::disconnect called, calls DcpProducer::closeAllStreams again
   - which calls ActiveStream::setDead; attempts to dereference null
   backfillMgr ptr (if in backfilling state).

This is a similar issue to MB-37702 and MB-38521 - the difference
being exactly when the streamRequest occurs relative to
closeAllStreams.

The first attempt to fix this issue was by changing
DcpProducer::disconnect from a atomic bool to a
folly::Synchronized<bool>, which can be locked for a dduration. We
could then acquire a read lock on DcpProducer::disconnect in the
prologue of DcpProducer::streamRequest, holding it until stream is
created to avoid the DcpProducer changing state partway through.

However, that introduced various lock inversions which were not
straightforward to solve - and addressing the first few significantly
increased the complexity of the fix which is undesirable this close to
7.0.

As such, the chosen solution is much simpler - check backfillMgr is
non-null before dereferencing it. In the case where a DcpProducer is
disconnecting and closing all it's streams, any backfillMgr buffer
updates are no lonnger relevent, so this is a valid solution.

Change-Id: Ide753264207d6ee8e7667358a36fb08ebc9ef11a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150259
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>