MB-45347: Backfill always reads the full payload for System Events
NO_VALUE must not apply to SystemEvent. That is because System Events
stores in the value information that is necessary for making the DCP
message streamed to the DCP client.

While that was already the case for in-memory streams, at backfill we
missed to pick up the value for NO_VALUE streams. That result is that we
throw and crash when we try to make the DCP message from a partial
payload in the ActiveStream::backfillReceived path.

Change-Id: I153533cd8eb3e524ef3b8aff12c73e73e059b1fd
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>